[external/boringssl] Add rules for pure-capabilities and make it work
This commit makes it so the SSL libraries work properly for CHERI
binaries.

In particular, it fixes some alignment issues which lead to the
library either not compiling (include/openssl/aead.h) or throwing a
SIGBUS error when loaded (src/crypto/mem.c).
An extra rule was added in Android.bp which makes it so Morello
doesn't use assembly when compiling, due to the possible issues
that using it may cause.

Signed-off-by: Lorenzo Carletti <lorenzo.carletti98@gmail.com>